refactor: add comments and simplify code
