+++ act.ts
@@ -1,4 +1,5 @@
-import { writeFileSec, WriteFileSecOptions } from "./utils.ts";
+import { log } from "./dev_deps.ts";
+import { writeFileSec } from "./utils.ts";
 import {
   actions,
   defaultModuleContent,
@@ -6,117 +7,15 @@ import {
   defaultTestModuleContent,
   Settings,
 } from "./settings.ts";
-import { asciiDeno } from "./ascii.ts";
-import { log } from "./dev_deps.ts";
 
-export async function act(settings: Settings) {
-  const path = settings.name;
-  if (path !== "." && path !== "./") {
-    await Deno.mkdir(path, { recursive: true });
-  }
+export const fns = {
+  writeFileSec,
+  initGit,
+};
 
-  if (settings.configOnly) {
-    await funcs.addModuleFile(
-      `${path}/deno.json`,
-      settings.configContent,
-      { force: settings.force },
-    );
-    if (settings.ascii) drawDeno();
-    return;
-  }
-
-  if (settings.js) settings.extension = "js";
-
-  if (settings.config) {
-    await funcs.addModuleFile(
-      `${path}/deno.json`,
-      settings.configContent,
-      { force: settings.force },
-    );
-  }
-
-  if (settings.importMap) {
-    await funcs.addModuleFile(
-      `${path}/import_map.json`,
-      settings.mapContent,
-      { force: settings.force },
-    );
-  }
-
-  await funcs.addModuleFile(
-    `${path}/${settings.entrypoint}.${settings.extension}`,
-    defaultModuleContent,
-    { force: settings.force },
-  );
-
-  await funcs.addModuleFile(
-    `${path}/${settings.depsEntrypoint}.${settings.extension}`,
-    defaultModuleContent,
-    { force: settings.force },
-  );
-
-  if (settings.tdd) {
-    const testModBytes = new TextEncoder().encode(
-      new TextDecoder()
-        .decode(defaultTestModuleContent)
-        .replace(
-          /\{\{extension\}\}/,
-          settings.extension,
-        ),
-    );
-
-    await funcs.addModuleFile(
-      `${path}/${settings.entrypoint}.test.${settings.extension}`,
-      testModBytes,
-      { force: settings.force },
-    );
-
-    await funcs.addModuleFile(
-      `${path}/${settings.devDepsEntrypoint}.${settings.extension}`,
-      defaultTestImportContent,
-      { force: settings.force },
-    );
-  } else {
-    await funcs.addModuleFile(
-      `${path}/${settings.devDepsEntrypoint}.${settings.extension}`,
-      defaultModuleContent,
-      { force: settings.force },
-    );
-  }
-
-  if (settings.ci) {
-    const actionsPath = `${path}/.github/workflows`;
-    await Deno.mkdir(actionsPath, { recursive: true });
-
-    await funcs.addModuleFile(
-      `${actionsPath}/build.yaml`,
-      actions,
-    );
-  }
-
-  if (settings.git) {
-    await funcs.initGit(path);
-  }
-
-  await funcs.addModuleFile(
-    `${path}/${settings.gitignore}`,
-    settings.gitignoreContent,
-    { force: settings.force },
-  );
-
-  if (settings.ascii) {
-    drawDeno();
-  }
-}
-
-async function addModuleFile(
-  filename: string,
-  content: Uint8Array,
-  options?: WriteFileSecOptions,
-) {
-  await writeFileSec(filename, content, options);
-}
-
+/**
+ * Responsible for calling `git init` in a directory. The function wrap is mainly to improve testability.
+ */
 async function initGit(path: string) {
   try {
     const cmd = Deno.run({
@@ -135,18 +34,101 @@ async function initGit(path: string) {
   }
 }
 
-function drawDeno() {
-  const lines = asciiDeno.split(/\n/);
-  let i = 0;
-  for (const line of lines) {
-    i += 50;
-    setTimeout(() => {
-      console.log(line);
-    }, 250 + i);
+/**
+ * This function is responsible for creating the module's contents using user-provided settings or defaults.
+ */
+export async function act(settings: Settings) {
+  console.log(settings);
+  const path = settings.name;
+
+  if (path !== "." && path !== "./") {
+    await Deno.mkdir(path, { recursive: true });
+  }
+
+  if (settings.configOnly) {
+    return await fns.writeFileSec(
+      `${path}/deno.json`,
+      settings.configContent,
+      { force: settings.force },
+    );
+  }
+
+  if (settings.ci) {
+    const actionsPath = `${path}/.github/workflows`;
+    await Deno.mkdir(actionsPath, { recursive: true });
+
+    await fns.writeFileSec(
+      `${actionsPath}/build.yaml`,
+      actions,
+    );
+  }
+
+  if (settings.config) {
+    await fns.writeFileSec(
+      `${path}/deno.json`,
+      settings.configContent,
+      { force: settings.force },
+    );
+  }
+
+  if (settings.js) settings.extension = "js";
+
+  await fns.writeFileSec(
+    `${path}/${settings.gitignore}`,
+    settings.gitignoreContent,
+    { force: settings.force },
+  );
+
+  if (settings.importMap) {
+    await fns.writeFileSec(
+      `${path}/import_map.json`,
+      settings.mapContent,
+      { force: settings.force },
+    );
+  }
+
+  await fns.writeFileSec(
+    `${path}/${settings.entrypoint}.${settings.extension}`,
+    defaultModuleContent,
+    { force: settings.force },
+  );
+
+  await fns.writeFileSec(
+    `${path}/${settings.depsEntrypoint}.${settings.extension}`,
+    defaultModuleContent,
+    { force: settings.force },
+  );
+
+  if (settings.tdd) {
+    const testModBytes = new TextEncoder().encode(
+      new TextDecoder()
+        .decode(defaultTestModuleContent)
+        .replace(
+          /\{\{extension\}\}/,
+          settings.extension,
+        )
+    );
+
+    await fns.writeFileSec(
+      `${path}/${settings.entrypoint}.test.${settings.extension}`,
+      testModBytes,
+      { force: settings.force },
+    );
+
+    await fns.writeFileSec(
+      `${path}/${settings.devDepsEntrypoint}.${settings.extension}`,
+      defaultTestImportContent,
+      { force: settings.force },
+    );
+  } else {
+    await fns.writeFileSec(
+      `${path}/${settings.devDepsEntrypoint}.${settings.extension}`,
+      defaultModuleContent,
+      { force: settings.force },
+    );
+  }
+
+  if (settings.git) {
+    await fns.initGit(path);
   }
 }
-
-export const funcs = {
-  addModuleFile,
-  initGit,
-};
