+++ validate_options.ts
@@ -9,7 +9,7 @@ export function validateOptions(options: Settings) {
       options.importMap || options.js)
   ) {
     throw new Error(
-      "Cannot use --config-only in combination with other options other than --name.",
+      "Cannot use '--config-only' with other options except '--name.'",
     );
   }
 }
